Fix a few x64 warnings

+++ src/core/indexing.cpp
@@ -268,7 +268,7 @@ int FFMS_Track::ClosestFrameFromPTS(int64_t PTS) const {
 	iterator Pos = std::lower_bound(begin(), end(), F, PTSComparison);
 	if (Pos == end())
 		return size() - 1;
-	int Frame = std::distance(begin(), Pos);
+	size_t Frame = std::distance(begin(), Pos);
 	if (Pos == begin() || FFABS(Pos->PTS - PTS) <= FFABS((Pos - 1)->PTS - PTS))
 		return Frame;
 	return Frame - 1;
